done TripIdsRepository, and started working on UsersRepository, already 600lines in tripsrepository )
